requirements.txt
#
# Python dependencies for the Unified Security Event Pipeline (M2 baseline)
#
# Core big-data / processing stack
pyspark

# Data handling and utilities
pandas
pyarrow
requests
python-dotenv
parquet-tools

duckdb

# AWS (S3 upload script + CLI)
boto3
awscli